requirements.txt
fastapi
uvicorn
pydantic
python-multipart
google-genai
groq
python-dotenv